requirements.txt
torch
numpy
matplotlib
hydra-core
rtree
tqdm
scikit-learn
tensorboard